chore(internal): improve performance of telemetry dependency reporting (#9515)
This change:

- Removes the usage of `TelemetryModuleWatchdog`
- A strategy of diffing `sys.modules.keys()` is now used instead of
watching for imports.
- Removes tracking of module origin
- Since we are only interested in distribution name + version we don't
need to keep track of all imported module origins, and then look up
distributions from that. We can use import names to directly look up the
distribution they are apart of.
- Add support for namespace packages
- Added the use of `importlib_metdata.packages_distributions()` to
ensure we are able to resolve namespace packages to their actual
distribution name.
- We only add `importlib_metadata` for Python <3.8, but
`packages_distributions` was added in 3.10, the footprint is fairly
small, so it is just vendored for the cases when we don't have it
available (keeps `importlib_metadata` requirement fairly broad as well)


The main change here is to have the TelemetryWriter keep track of which
`sys.modules.keys()` we have already resolved and sent information
about. Then on each periodic run, compare current `sys.modules.keys()`
against the cache, and send any new modules.

We also are no longer using origins to try and resolve the distributions
for a given module, instead we use
`importlib_metadata.distribution(name)`, if that fails to resolve, then
we will check to see if that import name is mapped to any namespace
packages, which we then try grab the distributions for that. (this means
any given module name could resolve to multiple distributions to report
about).

I have really struggled to get a solid benchmark/performance measurement
going, the only details I have are the internal app I was testing
against went from 100s to about 2s of overhead added for dependency
resolution, and the test suite itself still validates that the previous
behavior is occurring.


## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [ ] Title is accurate
- [ ] All changes are related to the pull request's stated goal
- [ ] Description motivates each change
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [ ] Testing strategy adequately addresses listed risks
- [ ] Change is maintainable (easy to change, telemetry, documentation)
- [ ] Release note makes sense to a user of the library
- [ ] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [ ] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)