fix(llmobs): don't default session id to trace id for llm obs span events (#10332)
We are defaulting `session_id` to be equal to `trace_id` in cases where
it was not set.

`session_id` is optional - we shouldn't be defaulting it to anything.
The front-end assumes that if `session_id` is present, it will toggle on
the session navigation view. We don't want the session view to be
toggled on for traces or ml apps where session views don't make sense.

**how will this effect customers?**

1. This will break code paths for customers with code-paths that rely
`session_id` always existing on the underlying span. Though this is very
unlikely, as this change only applies to customers who never set
`session_id` in the first place (meaning they may not know what
session_id is, and thus do not rely on it for anything else)

2. customers will see UI changes where the session navigation view
disappears (or potentially any graphs that rely on `session_id`, not
this is unlikely due to the reasoning in point `1`). These UI changes
are for the better, since the existing session navigation view is either
a) useless (if session_id = trace_id, there is only one trace belonging
to each session) or a mis-use / mis-representation of what `session_id`
semantically means.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: lievan <evan.li@datadoqhq.com>
Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>
(cherry picked from commit 27335e302f39caf9dddac3f3d67c17d695b25ab4)