feat(llmobs): support tagging integration-generated spans (#10638)
This PR introduces `AnnotationContext` and `AnnotationContextAsync`,
which are context managers that allow users to modify integration
generated spans.

Create an `AnnotationContext` object by calling
LLMObs.annotation_context(). On entering this context, a function is
registered on span startup that modifies the span with data passed into
annotation_context()

This PR only exposes a `tags` argument for `annotation_context`. This is
a good candidate as a first argument because,
1. customers have asked for it
2. span kind agnostic
3. there's simple merging logic for tags

We also modify the `annotate` function to allow for tagging non-IO span
data which does not depend on span kind so annotate can be re-used
within the annotation context

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: lievan <evan.li@datadoqhq.com>
Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>