chore(ci_visibility): use two-pass procedure for code instrumentation in Python 3.10 (#10669)
As my project for Innovation Week, I experimented with a more efficient
way to do code coverage instrumentation in Python 3.10 (the idea is to
apply this logic to the other versions in the future; I did it on Python
3.10 because it's what we use in dogweb).

The idea is to make all jumps fixed-size by emitting a fixed number of
`EXTENDED_ARG` instructions before each jump, regardless of whether they
are strictly necessary or not. This allows the jump targets to be
rewritten in-place, without worrying that a new jump target might
increase the number of `EXTENDED_ARG`s needed, which in turn would
change the offsets of following instructions, which would change the
jump targets, which might increase the number of `EXTENDED_ARGS`, and so
on.

In this way, the instrumentation can be written as a two-pass procedure:
first the instrumented bytecode is emitted with placeholders for the
jump targets, and then the placeholders are patched with the final
offsets of the jump targets. The algorithm is further explained in a
long comment in the code.

As a result of these changes, the time taken to instrument code while
running the Dogweb delancie tests drops from 7.8 seconds to around 2.5
seconds (about 3x as fast).

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)