chore(ci): remove lib-init image tests from this repo in favor of images built in system-tests (#10746)
This PR removes lib injection test images from this repo in favor of the
ones in `system-tests`. The images in this repository were not used
since the migration to `one-pipeline`. The transition from Github
actions to Gitlab for kubernetes lib injection tests mean none of the
injection tests in this repo ran.

With tests in both locations there is lots of potential for confusion.

To capture the changes that have been mistakenly applied to
`dd-trace-py` images, there is a sister PR in system-tests:
https://github.com/DataDog/system-tests/pull/3089

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
# Conflicts:
#	.gitlab-ci.yml