chore(di): guard against mutable module container
When disovering functions on module import, we look at the module as a
container of function-like objects. If the module is partially loaded,
its `__dict__` might mutate. We ensure to iterate over a copy of the
module's `__dict__` when iterating over it. We expect the original
module to mutate dynamically at runtime, so by the time we make a copy
of its underlying lookup dict, we are able to detected at the very least
everything that corresponds to the actual source code.