fix(elasticsearch): set tags on span, regardless of sampling
The Datadog ElasticSearch integration avoids setting metadata on spans
when those spans are marked to be dropped.  A recent change was added
for sampling to lazily determine whether a span is sampled - which means
the sample decision is much too late for this code (see
https://github.com/DataDog/dd-trace-py/pull/8308).

Prior to this fix, some elasticsearch query spans incorrectly failed to
set metadata based upon sampling rules, and thus incorrectly
participated in metrics.

After this fix, the metadata is set regardless of sampling decisions.