fix(llmobs): refactor annotation context implementation (#10976)
This PR fixes two issues with annotation contexts
1. Nested annotations were applied in a random order
2. Annotations polluted across all trace contexts because they were
registered as a global `on_span_start` hook

I put it all in one PR since it's a single refactor of the underlying
implementation for annotation contexts, but let me know if you would
prefer this to be separated into two PR's -

**Issue 1: Nesting**
Currently, annotation contexts register a single `annotate` function to
be called with specific arguments on span start when we the context
manager is entered.

If multiple annotation contexts are nested, the order in which the
annotations happen are non-deterministic. This is because span start
function hooks are stored as a set.

Instead of registering each annotate function as a separate hook, we
only register one annotation function on span start that iterates
through a list of annotation arguments.

Entering/exiting the annotation context enqueues/dequeues arguments from
the list of annotation arguments stored in the LLMObs instance.

This ensures that annotations are applied in the order in which
annotation contexts are entered.

**Issue 2: Cross trace context pollution**
Annotations are applied as a global function hook on span start, meaning
they apply to all spans across all trace contexts.

We update this so that we store each annotation invocation with an
`annotation_context_id` representing the current trace context.
1. If a trace has not started yet and there is no active context, we
create and activate a "dummy" context containing `annotation_context_id`
in the baggage
2. If there is already a currently active context, we either extract the
existing `annotation_context_id` from the baggage or add a new one

When we apply the annotations, we fetch the `annotation_context_id` from
the current trace context and check that it matches the
`annotation_context_id` stored with each annotation.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: lievan <evan.li@datadoqhq.com>
(cherry picked from commit 402cfb75afb06343674b65966bc6e50f355c7992)