chore(core): core update for span management (#11012)
Simplify the core "call_key" mechanism with a simple "span" attribute
for the core context (based on a new _innerspan attribute).

- `ctx.set_item(ctx.get_item("call_key"), span)` become `ctx.span =
span`
- `span = ctx.get_item(ctx["call_key"])` become `span = ctx.span`


Add a "span_key" mechanism to automatically add the span to the context
data for simple access in case we need to access a span by name on a non
current context.


* Technical Notes *
I wanted to ensure that the inner span for a context is only set once.
This is not the case, so I removed the condition. In the previous
version, different spans could be assigned to the call_key during
context life. As the logic was preserved in this PR, this is also the
case with ctx.span.


APPSEC-55053

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)