chore(ci): fix django snapshot timeouts (#11138)
This PR improves and fixes issues in the Django tests with the following
changes:
- Share the PATH and PYTHONPATH with the subprocess: When a script runs
in a subprocess, there are times in the CI or locally when it's assigned
a different path than expected. Even worse, we've seen scripts that
worked for months suddenly stop working because of this. With this
function, we always set the path to ensure consistent results both
locally and across different CI environments
- Replace the `ddtrace-run` command with `python -m
ddtrace.commands.ddtrace_run` to ensure that the local library is
executed, rather than a potentially installed version
- Displays the stdout and stderr if there was an issue in the subprocess

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>